refactor(saga): split handler execution internals
