+++ .changeset/red-sagas-split.md
@@ -1,0 +1,5 @@
+---
+"@redemeine/saga": patch
+---
+
+Refactor saga handler execution internals into focused modules without changing public API or runtime behavior.
